feat: refactor chain config, add new chains
